test: extract property filter test cases to a dedicated test class
